🐛 [video-creator] Prevent screenshots before simulation update
Fixes #132: slow simulation updates lead to premature screenshots.

This does *not* use the property `onSimulationUpdatesAppliedSpy`;
instead uses `observeEvent("change")`, though it may be possible
for the user to make an edit between dispatching the simulation step
and taking the screenshot, causing the screenshot to be taken before
the dispatched step is applied.

Another issue with `observeEvent("change")` is that a simulation that
does not change the graph state would not trigger properly, instead
pausing the capture in an unintended way. However, no one should
capture ineffective simulations, so this should be fine to leave in.