Remove need for require() on main frame
Now, reference everything from `Desmos.Private.Fragile`, and fall back to the current `require()` until Desmos  has switched over.

The worker `require()`s are not converted yet; the `Desmos` API object is not available in the worker. I'll have to find another way to expose the functions, or get around the need for them.