Add confirmation for GLesmos lines
This only applies to GLesmos inequalities like `sin(xy)>0`, and not equalities like `sin(xy)=0` since equalities could not be plotted before the outlines update.

The primary purpose of this change is to ensure that graphs created pre-outlines would not suddenly get much slower post-outlines. Adding a new confirmation check avoids the stateful mess of the alternative (version migration, changing the interpretation of the lines property, etc.).

A secondary purpose to this update is to show a little message before people enable lines. This has minor utility and only appears for inequalities.