Avoid annoying overwriting (most cases)
Fixes #393

Table columns and image use focus.

Regressions only update the parameters `#{ ... }` block. The regressions can still be improved (what happens if the user thinks there's a `#{` block, but syntax error? Then there's a new `#{` block, bad).

Update to @codemirror/view is needed because it fixes a bug with focus.

https://discuss.codemirror.net/t/how-to-add-decoration-without-focusing-the-editor-on-dispatch/6283/2

https://github.com/codemirror/view/commit/f49f2eea1ae1b81c8e943b76a57ca787bfd46bf5