Remove home link disable on panic popover
I presume this change was made a while ago when the panic button was on
top of Desmos' home link. I was checking the git history of this line,
and it seems the last changed about a year ago in May, so I think it
lines up. Either way the purpose of this change is because currently the
link is unclickable until the popover is open, but then it is blocked,
so it renders the home link unusable.