Fix missing styling on blue buttons
The _overrides.less file is not supposed to be edited manually, but I
tried to regenerate from the raw css of Desmos and not only the
resulting file was massive compared to its current size but it had tons
of errors with unmatched brackets. I could not fix that so I opted for a
manual edit.