Fix screeshot scaling issue
The changes in index.ts and CaptureMethod.tsx fix the issue that some
screenshots with smaller width and or height.

The changes in capture.ts are a misplaced comment and that for some
reason "contain" was not the default behavior for screenshot mode when
the regular API was used (as opposed to fast screenshot).